9993-net_essedma_remove_default_vlan_tagging.patch
ipq40xx: remove default VLAN tagging for ESSEDMA

This patch removes the default VLAN tagging, which is used by this switch to pretend to have 2 CPU ports. Actually, the switch only have one CPU port, at least on some devices, and for them it's better to use the kernel's VLAN tagging.

This way, users can use VLAN as expected and the VLANs 1 and 2 become free (not longer reserved by the switch efforts to pretend it's a two MAC device). This, also, fully avoids the double-tagging when actually using a VLAN either on LAN or WAN.

Signed-off-by: Will Owen <will.j.owen@gmx.com>